requirements.txt
Flask==3.0.0
twilio==8.10.0
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.1
requests==2.31.0
Werkzeug==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
groq==0.9.0
httpx==0.27.2
tensorflow
pillow
numpy
tensorflow-macos
tensorflow-metal
